@@ -5,19 +5,21 @@
 // or any plugin's vendor/assets/javascripts directory can be referenced here using a relative path.
 //
 // It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
-// compiled file.
+// compiled file. JavaScript code in this file should be added after the last require_* statement.
 //
 // Read Sprockets README (https://github.com/rails/sprockets#sprockets-directives) for details
 // about supported directives.
 //
 //= require jquery
-//= require jquery.turbolinks
 //= require jquery_ujs
 //= require bootstrap
-//= require react
-//= require react_ujs
-//= require components
-//= require_tree .
 //= require turbolinks
+//= require_tree .
 //= stub rails_admin/custom/ui.js
 //= stub 'ie'
+
+$(document).on('turbolinks:load', init);
+
+function init() {
+  console.log('Code ready to rocks !');
+}
