@@ -3,8 +3,7 @@ module RailsAdmin::SeoAdmin
 
   included do
     rails_admin do
-      navigation_label 'SEO'
-      navigation_icon 'fa fa-info-circle'
+      navigation_icon 'fa fa-google'
       parent Content
 
       edit do
