@@ -1,17 +1,15 @@
 Rails.application.routes.draw do
+  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+
   # USERS & ADMINISTRATION
   devise_for :admins
-  devise_for :users
   mount RailsAdmin::Engine => '/admin', as: 'rails_admin'
-
-  # HOME
-  root to: 'home#index'
-
-  # PAGE
-  get '/pages/:id' => 'pages#show', as: 'page'
 
   # FROALA (WYSIWYG)
   post '/froala_upload' => 'froala#upload'
   post '/froala_manage' => 'froala#manage'
   delete '/froala_delete' => 'froala#delete'
+
+  # HOME / root
+  root to: 'site#index'
 end
